facs: Add a check for empty BIOS memory map info list
FWTS compliance test suite tries to gather the BIOS memory map list
info from the sysfs file called "/sys/firmware/memmap", in case of
failure it retry to get the same from the klog(kernel log). If it
fails here also then we should not check that allocated memory for
FACS structure is reserved or not, because we don't have memory map
info to compare.

This patch fixes this issue by adding a check for empty memory_map_list
before comparing.

Signed-off-by: Rameshwar Prasad Sahu <rsahu@apm.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>