lib: fwts_set: fix API for fwts_set, add fwts_set_int
The fwts_set API has the args the wrong way around compared to
fwts_get, so swap these; fortunately it is not being used yet
in fwts, so this won't break anything.  Also add fwts_set_int
to set integer values in /sys or /proc files.  Finally, add
error checking in case the writes to the file fails.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>