efi_runtime: build dkms for kernel that does not have efi_test
A version of efi_runtime, efi_test, has been sent to upstream since
kernel 4.9, and all supported Ubuntu kernels, except 4.4, haven efi_test
built as a module by default. As a result, the dkms is only required for
kernel 4.4. The also fixes the problem that dkms fails to be installed
on new hardware that has secure boot enabled by default.

Signed-off-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.coACK