nfit: scan SMBIOS Management Information structure
Perform a scan of the ACPI NFIT SMBIOS Management Information structure,
which contains extra SMBIOS structures that were not present at boot
and included in the real SMBIOS table (e.g., for hot-plugged Memory
Devices (type 17)).  Print the header information, including the
number of string bytes, and check that the lengths are plausible.

This could be enhanced to share functions in dmicheck.c like
dmi_scan_tables and dmicheck_entry to do more detailed checks.

Signed-off-by: Robert Elliott <elliott@hpe.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>