fwts_coreboot_cbmem: fix off-by-one error on last char of buffer
The coreboot_log has been allocated a size of console->size + 1,
however, the end of string terminator '\0' is being written one
character passed the end of the buffer because of an off-by-one
error. Fix this.

Also insert some spaces as per the fwts coding style.

Detected by CoverityScan, CID#1394472 ("Out-of-bounds write")

Fixes: c9bf07f25d13 ("fwts_coreboot.c: add cbmem console parser")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>