lib: fwts_acpi_tables: unconstify data and don't strcpy ACPI table ids
Although it seems a good idea to make the ACPI table data const, we
may need to modify it when patching pointers if we make cached copies
of the data, so unconstify it.

Also, don't strncpy ACPI table IDs, instead memcpy them as they are
not really strings with terminating zero bytes.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>