DMI: Fix the way to load SMBIOS 3.x tables via sysfs
In SMBIOS 3.x specs, the meaning of "Structure table maximum size"
field in SMBIOS Entry Point Structure is "Maximum size of SMBIOS
Structure Table, pointed to by the Structure Table Address, in bytes.
The actual size is guaranteed to be less or equal to the maximum size."

Thus we cannot assume the whole size of SMBIOS segment equals to the
value of entry->struct_table_length.

This fix the DMI test failure in some platforms when the real size of
SMBIOS segment are smaller than the value in "Structure table maximum
size" field

Signed-off-by: Ge Song <ge.song@hxt-semitech.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>