kernelscan: add -k option to specify klog json filename
This allows one to specify the location of the klog json file. Currently
this is hard coded to be /usr/share/fwts/klog.json, but this new option
allows one to specify the full path. This allows us to use the json file
in the fwts src rather than the installed version.

E.g.

./kernelscan.sh -k ~/repos/fwts/data/klog.json  ~/repos/linux

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>