lib: fw_pm_method: move the sleep t_start time out of logind_on_signal (LP: #1809096)
Some systems that t_start time will be recorded after resume, that causes false
alarm, the duration be 0. This is because storing t_start is on
logind_on_signal(), waiting for "PrepareForSleep" signal, which hasn't be
triggled, but the system has already slept. Move the t_start recording right
before doing the s3/s4, also move the kernel logging, so that the logs can be
in the right place and have right time record on the dmesg, which are used to
calculate the suspend time.

Signed-off-by: Robert Liu <robert.liu@canonical.com>
Signed-off-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>