fwts_acpica: don't add in RSDP or null table entries to XSDT or RSDT (LP: #1829167)
BugLink: https://bugs.launchpad.net/fwts/+bug/1829167

Adding in the RSDP table to the XSDT and RSDT is illegal. It also
breaks ACPICA as this tries to checksum the RSDP and as it's a non-standard
table the length field is offset differently from most tables causing
the checksum to segfault.  Also don't in null tables, this does not make
sense.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>