uefi: uefirttime: make status static, cleans up cppcheck warnings
Using an auto-variable (declared on the stack) for the status
is problematic because we return the address of this variable
back to the callee via a function parameter. Thus there is a
potential of the callee referencing the now dead stack variable.
Although this is not actually referenced later, it's worth just
cleaning this up by making the variable static so we avoid any
potential risk in the future.  Cleans up 3 cppcheck warnings
of the kind:

"(error) Address of local auto-variable assigned to a function parameter."

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>