live-image: overwrite pre-defined tests by oem-tests.lst
This adds a feature to check /fwts/oem-tests.lst and uses its contents as
tests to be run. The oem-tests.lst is to be put to "fwts-result"
partition of the fwts-live USB disk. This allows fwts-live users to
customise fwts-live for their needs without creating special fwts-live
images.

The oem-tests.lst has the following format:

TESTS=list of tests to be executed
POST=poweroff|reboot

For example:
TESTS=mcfg aspm
POST=poweroff

Signed-off-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>