debian/tests: only run dkms test when dkms module is run
In the case where the efi test driver is not built by dkms
then we should basically skip the dkms test.  To do this,
add a test control file that only runs the dkms-autopkgtest
if the dkms driver has been built.

This commit is requred by the earlier commit "efi_runtime: don't
build dkms module for kernels 4.8.0 and above" to work on older
and current kernels.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>