lib: fwts_multiproc: add sanity check to data->size to clean up static analysis warning
Static analysis warns that the data->size is not validated so add a simple
sanity check on the size.  This is a moot point as we know the size is never
going to be huge and we have to trust the given table size, but this at least
sanity checks things and keeps static analyzers happy.

Addresses-Coverity: ("Untrusted value as argument")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>