lib: fwts_modprobe: add helper to build module path name
The current method of building a module pathname may cause truncation
and gcc-10 generates build errors because of this. Add a helper that
builds the module path that uses the BSD strlcat for safe string
concatenation with known (and ignored) truncation since we know that
the module path name can't exceed PATH_MAX size.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>