Replace json-c library with minimal replacement
FWTS uses a subset of json-c functionality, so replace it with
our own implementation that supports this subset. This also
removes the dependency of two flavours of json libraries that
fwts has been using and also allows us to perform deeper memory
allocation and leaking tracking with static analysis tools.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>