s3: fix segmentation fault when PM_S2IDLE_SLP_S0 is not available (LP: #1894229)
Currently if PM_S2IDLE_SLP_S0 is not available then atoi() is called
on a NULL pointer causing a segmentation fault.  If it is available
there string returned is leaked.  Fix these two bugs by adding a helper
function to read the PM_S2IDLE_SLP_S0 file and return 0 if it is not
possible or the residency value if it is available.

Fixes: 69c18bab0cc0 ("s3: check slp_s0_residency_usec for s2idle")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>