acpi: fan: skip over fan info when the type field is NULL (LP: #1906540)
It is possible for a read error on the thermal type field to end up
with a null type field. (This can occur when a snap is not installed
in --devmode and cannot access thermal control data).  This causes a
segmentation fault when dereferencing the pointer. Add null checks to
skip over such empty data fields.

Fixes: 10b9561f46f8 ("apci: fan: use newer thermal interface, test fans on loaded CPU (LP: #831905)")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>