acpi: lib: correct revision of the fake XSDT table
According to ACPI 6.3 specification, section 5.2.8 Extended System
Description Table (XSDT), the 'Revision' field of XSDT table is 1,
not 2. So, the current fake XSDT table generated by fwts causes the test
case ACPITableBadRevision failed due to the revision is not matched with
ACPI 6.3.

This patch changes the revision of the fake XSDT table from 2 to 1 as
this revision is compatible both with ACPI 6.3 and older versions.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>