uefi: clean up a cppcheck style warning
Cppcheck is complaining because the structure capsuleheader
is being referenced but it is never used. It is probably a
good idea to zero this structure to silence this warning and
to ensure the referenced data does not contain garbage.

Cleans cppcheck warning:
src/uefi/uefirtmisc/uefirtmisc.c:102:21: style: Unused
variable: capsuleheader [unusedVariable]
 EFI_CAPSULE_HEADER capsuleheader

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>