cpu: cpufreq: initialize online_cpus in for loop
In the case where macro FWTS_ARCH_INTEL is not defined the
function returns early and the initialization of online_cpus
throws a cppcheck style warning because it has been initialized
but has not been used.  Move the online_cpus initialization to
the start of the for-loop to clear this warning.

Cleans up cppcheck warning:
src/cpu/cpufreq/cpufreq.c:398:21: style: Variable 'online_cpus'
is assigned a value that is never used. [unreadVariable]
 int i, online_cpus = 0;

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>