lib: fwts_acpi_tables: update fwts_get_acpi_version (mantis: 2105)
ACPI 6.4 changes FADT minor format to include errata as followings:

Bits 0-3 - The low order bits correspond to the minor
version of the specification version. For instance, ACPI
6.3 has a major version of 6, and a minor version of 3.

Bits 4-7 - The high order bits correspond to the version
of the ACPI Specification errata this table complies
with. A value of 0 means that it complies with the base
version of the current specification. A value of 1 means
this is compatible with Errata A, 2 would be
compatible with Errata B, and so on.

Signed-off-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>