fwts_devicetree: clean up pedantic static analysis warning
Coverity is warning that prop_len may be uninitialized and prop_buf
could be null in the fdt_getprop call. Not sure of these are false
positives, but clean up the warning by initializing prop_len to zero
and call strcmp on prop_buf only if it is not null.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>