lib: fwts_coreboot_cbmem.c: clean up static analysis warning
Static analysis with clang scan-build is warning about cbmem_console_addr
being not initialized and being passed to a map_memory call. Avoid this
by explicitly initializing it to (off_t)-1 and checking for this not being
changed in a for-loop rather than checking if the loop reached the end
without breaking.  This simplifies things for static analyzers.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>