lib: fwts_acpi_tables: fix the passed reserved value
BugLink: https://bugs.launchpad.net/fwts/+bug/2000752

The function fwts_acpi_reserved_type verify the type to see if in reserved,
which compares with the value with >=
if (value < min || value >= reserved) {

So it shouldn't pass the reserved value with RESERVED - 1.
Also modify the print out maximum value for reserved.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>