auto-packager: mkpackage: add build depend dh-dkms for lunar
lunar build fail, when dh-dkms no installed, add the dh-dkms to
Build-Depends when prepare build for lunar.
dh-dkms is only on lunar and kinetic, not on other releases, so
not adding directly to debian/control for all releases in case we don't have
build error in those releases which no dh-dkms package.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>