acpi: button: remove check fixed hardware & control method power buttons
Buglink: https://bugs.launchpad.net/fwts/+bug/2017737

In ACPI spec:

A zero indicates the power button is handled as a fixed feature programming
model; a one indicates the power button is handled as a control method device.
If the system does not have a power button, this value would be “1” and no
power button device would be present. Independent of the value of this field,
the presence of a power button device in the namespace indicates to OSPM that
the power button is handled as a control method device.

There is no linkage that having a PNP0C0C device and that being zero is a
problem.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>