acpi/wmi: Check if the mandatory block query method exists
According to the ACPI WMI documentation, WMI data blocks
need to provide a block query method (WQxx). On several
machines (like the Dell Inspiron 3505), some WMI data blocks
do not provide such a method, which for example causes
the linux kernel to log a warning.
Add a matching test to allow for catching this sort of
issue with the fwts. Also update the wmi regression tests.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>
Acked-by: Ivan Hu <ivan.hu@canonical.com>