acpi: fadt: remove the tests for expecting smi_cmd is zero
BugLink: https://bugs.launchpad.net/fwts/+bug/2032158

We should not ecpect smi_cmd to be zero if acpi_enable/acpi_disable is zero.
Checking the ACPI specification,
ACPI_ENABLE This field is reserved and must be zero on systems that do not
support Legacy Mode.
ACPI_DISABLE This field is reserved and must be zero on systems that do not
support Legacy Mode.
SMI_CMD This field is reserved and must be zero on system that does not support
System Management mode.
It doesn't force the SMI_COM shoud be zero if acpi_enable/acpi_disable is zero.

Some projects enable ACPI mode with DXE_DRIVER and doesn't switch back to
legacy mode once booted to OS. They support SMM as well.
So it got,
SMI_CMD is not zero
ACPI_ENABLE is zero
ACPI_DISABLE is zero

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>