tpmevlog: Correct boundary checking for PCR4 event check
The boundary checking looks when PCR 4 and `EV_EFI_BOOT_SERVICES_APPLICATION`
that the size is greater than `uefi_image_load_event`.

This test fails when loading a PE binary for a kernel image which is exactly
the size of `uefi_image_load_event`.

Furthermore, a kernel image may not have a device path specified, so don't
bail if that's zero size.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>