Upgrade MCFGMMIONotReserved to HIGH error
This impact from a BIOS with this problem is actually quite dramatic.
On a laptop with a dGPU that uses resizable BAR it will cause the dGPU
to stop working after runtime resume from D3cold or from s2idle.

To make it more obvious it must be fixed during BIOS development upgrade
the error to High.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>