acpi: acpipld: add tests to check _PLD methods only on the connectable ports
BugLink: https://bugs.launchpad.net/fwts/+bug/2055855

ACPI specification implies that a port is not connectable and it is assumed to
be not visible. Therefore a \_PLD descriptor is not required.

Some firmware have _PLD ACPI methods for all USB ports, including unusable
USB host ports that are not wired to any connector or internal device.
And these unusable ports return similar _PLD objects, with zeroes in their
grouptoken and position fields. This confuses the port peering code when
pairing USB2 and USB3 ports that are wired to the same connector. And it is
results in the linux kernel warning "usb: port power management may be
unreliable."
Add tests to check the PLD methods existencs on the connectable ports.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>