configure.ac: fix the autoreconf AC_PROG_LEX warning
BugLink: https://bugs.launchpad.net/fwts/+bug/2075488

autoreconf -ivf results in

autoreconf: running: /usr/bin/autoconf --force
configure.ac:7: warning: AC_PROG_LEX without either yywrap or noyywrap
is obsolete

According to https://www.gnu.org/software/autoconf/manual/autoconf-2.70/html_node/Particular-Programs.html
that changed in 2020:

Prior to Autoconf 2.70, AC_PROG_LEX did not take any arguments, and its
behavior was different from either of the above possibilities: it would
search for a library that defines yywrap, and would set LEXLIB to that
library if it finds one. However, if a library that defines this
function could not be found, LEXLIB would be left empty and LEX would
not be reset. This behavior was due to a bug, but several packages came
to depend on it, so AC_PROG_LEX still does this if neither the yywrap
nor the noyywrap option is given.

Usage of AC_PROG_LEX without choosing one of the yywrap or noyywrap
options is deprecated. It is usually better to use noyywrap and define
the yywrap function yourself, as this almost always renders the LEXLIB
unnecessary.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>