tpmevlog: Ensure the event log matches the actual TPM PCRs
Ultimately the TPM event log is used to be able to validate how the
current TPM PCRs values were arrived at. This is only useful if parsing
the event log results in the same values as the TPM actually has stored
in it. Add another minor test to the tpmevlog test that this is the
case.

Signed-off-by: Jonathan McDowell <noodles@meta.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>