acpi: uniqueid: fix the uniqueid test for the CID with package
https://bugs.launchpad.net/fwts/+bug/2091350

There are some false alarm failures due to the _CID with package,
even if the _UID is unique. This is becase _CID with package is not
handled, so cause hid_match is wrong. Add checking _CID with Package
type. Compare all items in the package, and if any item matches,
it is considered a match (as the OS might use the same one).
Then, further verify the _UID value.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>