fwts: skip the wake alarm suite if the feature is not implemented
The wake alarm may be not implemented on a non-x86 platform
thus the calls to the RTC alarm interface may legitimately
fail. Although the wake alarm test suite appears to antici-
pate this and skips the first test on a non-x86 platform if
fwts_wakealarm_get() fails the routine doesn't detect that
the wake alarm is not implemented thus the 1st test passes
and the remaining tests fail if it is the case. Here's an
excerpt from the FWTS results log obtained in such an envi-
ronmemnt:

 wakealarm: ACPI Wakealarm tests.
 ---------------------------------------------------------
 Test 1 of 5: Test existence of RTC with alarm interface.
 PASSED: Test 1, RTC with a RTC alarm ioctl() interface
 found.

 Test 2 of 5: Trigger wakealarm for 1 seconds in the future.
 Trigger wakealarm for 1 seconds in the future.
 Cannot enable alarm interrupts on Real Time Clock device
 /dev/rtc0.
 FAILED [MEDIUM] WakeAlarmNotTriggeredTest2: Test 2, RTC
 wakealarm did not trigger.

 Test 3 of 5: Test if wakealarm is fired.
 Cannot enable alarm interrupts on Real Time Clock device
 /dev/rtc0.
 FAILED [MEDIUM] WakeAlarmNotTriggeredTest3: Test 3, Failed
 to trigger and fire wakealarm.

 Test 4 of 5: Multiple wakealarm firing tests.
 Trigger wakealarm for 1 seconds in the future.
 Cannot enable alarm interrupts on Real Time Clock device
 /dev/rtc0.
 FAILED [MEDIUM] WakeAlarmNotTriggeredTest4: Test 4, Failed
 to trigger and fire wakealarm.

 Test 5 of 5: Reset wakealarm time.
 PASSED: Test 5, RTC wakealarm set.
 FAILED [MEDIUM] WakeAlarmNotResetTest5: Test 5, RTC wakealarm
 failed to be reset back to original time.

This happens because the routine fwts_framework_run_test() stops
performing the suite only if its test returns FWTS_ABORTED or its
init callback returns FWTS_SKIP while neither is the case here.

The patch fixes the issue by adjusting the routines wakealarm_
test[2-5]() to return FWTS_SKIP if the 1st test is skipped.

The wakealarm suite report looks as follows after the patch
is applied:

 wakealarm: ACPI Wakealarm tests.
 ------------------------------------------------------------
 Test 1 of 5: Test existence of RTC with alarm interface.
 The wake alarm is not supported.
 non-x86 devices sometimes do not have an RTC wake alarm that
 is normally controlled by the RTC alarm ioctl() interface.
 This interface does not exist, so the wake alarm tests will
 be skipped.

 Test 2 of 5: Trigger wakealarm for 1 seconds in the future.

 Test 3 of 5: Test if wakealarm is fired.

 Test 4 of 5: Multiple wakealarm firing tests.

 Test 5 of 5: Reset wakealarm time.

Signed-off-by: Al Korv <alkorv@posteo.uk>
Acked-by: Ivan Hu <ivan.hu@canonical.com>